pyproject.toml
[project]
name = "flask-api-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-pydantic>=0.13.0",
    "google-cloud-firestore>=2.20.1",
    "grpcio==1.46.3",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "mock-firestore>=0.11.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]
